requirements.txt
Flask==2.0.1
werkzeug==2.0.1
torch==2.0.1
torchvision==0.15.2
pillow
gunicorn
numpy==1.24.3